REFACTOR: Explicit tagged union of analysis types.
Per review by @olynch.
