+++ packages/frontend/src/analysis/index.ts
@@ -1,3 +1,4 @@
+export * from "./types";
 export * from "./model_graph";
 export * from "./stock_flow_diagram";
 export * from "./submodel_graphs";
