+++ packages/frontend/src/model/types.ts
@@ -1,10 +1,10 @@
-import type { Component } from "solid-js";
 import { uuidv7 } from "uuidv7";
 
 import { DblModel } from "catlog-wasm";
 import type { DblTheory, MorDecl, MorType, ObDecl, ObType } from "catlog-wasm";
+import type { ModelAnalysis } from "../analysis";
 import type { Notebook } from "../notebook";
-import type { Theory, TheoryId } from "../theory";
+import type { TheoryId } from "../theory";
 
 /** A model of a double theory in the form of notebook.
  */
@@ -19,7 +19,7 @@ export type ModelNotebook = {
     notebook: Notebook<ModelJudgment>;
 
     /** Analysis of the model, a separate notebook. */
-    analysis: Notebook<ModelAnalysis<unknown>>;
+    analysis: Notebook<ModelAnalysis>;
 };
 
 /** A judgment in the definition of a model.
@@ -77,42 +77,3 @@ export function catlogModel(theory: DblTheory, judgments: Array<ModelJudgment>):
     }
     return model;
 }
-
-/** Analysis of a model of a double theory.
-
-Such an analysis could be a visualization, a simulation, or a translation of the
-model into another format. Analyses can have their own content or state going
-beyond the data of the model, such as numerical parameters for a simulation.
- */
-export type ModelAnalysis<T> = {
-    /** Identifier of the analysis, unique relative to the theory. */
-    tag: string;
-
-    /** Content associated with the analysis (not the model). */
-    content: T;
-};
-
-/** Component that renders an analysis of a model. */
-export type ModelAnalysisComponent<T> = Component<ModelAnalysisProps<T>>;
-
-/** Props passed to an analysis of a model. */
-export type ModelAnalysisProps<T> = {
-    /** The model being analyzed. */
-    model: Array<ModelJudgment>;
-
-    /** The `catlog` representation of the model, if the model is valid. */
-    validatedModel: DblModel | null;
-
-    /** Theory that the model is of.
-
-    Some analyses are only applicable to a single theory but the theory is
-    passed regardless.
-     */
-    theory: Theory;
-
-    /** Content associated with the analysis itself. */
-    content: T;
-
-    /** Update content associated with the analysis. */
-    changeContent: (f: (content: T) => void) => void;
-};
